[ Mixed Tensor ] Enable FP32 unittest cases
This PR enables the FP32 unittest cases. It includes various fixes and
adding compiler preprocessor pragmas.

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

+++ nntrainer/tensor/tensor_pool.h
@@ -197,7 +197,7 @@ public:
   Tensor *view(const std::string &name, const std::string &reference,
                const TensorDim &dim,
                const std::vector<unsigned int> &exec_order,
-               TensorLifespan lifespan, const unsigned int offset = 0);
+               TensorLifespan lifespan, const size_t offset = 0);
 
   /**
    * @brief extend a tensor life as tensor is being shared.
